modify: apart nested type GameResult and Hands
